fix lỗi quyền xem,sửa xóa bài viết ngoài site

+++ modules/faq/funcs/list.php
@@ -11,10 +11,10 @@
 if (! defined('NV_IS_MOD_FAQ')) {
     die('Stop!!!');
 }
-if($module_setting['user_post']!=1) {
+if($module_setting['user_post']!=1 or empty($user_info['userid'])) {
 		Header("Location:"  . NV_BASE_SITEURL . "index.php?" . NV_LANG_VARIABLE . "=" . NV_LANG_DATA . "&" . NV_NAME_VARIABLE . "=" . $module_name);
 	                exit();
-	}
+}
 //List faq
 $listcats = array();
 $listcats[0] = array(
@@ -31,10 +31,6 @@ $per_page = 20;
 //List bài viết đã được duyệt
 $sql = "SELECT SQL_CALC_FOUND_ROWS * FROM " . NV_PREFIXLANG . "_" . $module_data . "";
 $base_url = NV_BASE_SITEURL . "index.php?" . NV_LANG_VARIABLE . "=" . NV_LANG_DATA . "&" . NV_NAME_VARIABLE . "=" . $module_name."&". NV_OP_VARIABLE . "=list";
-$where='';
-if($user_info['userid']!=0) {
-	$where .="userid=".$user_info['userid'];
-}
 if ($nv_Request->isset_request("catid", "get")) {
     $catid = $nv_Request->get_int('catid', 'get', 0);
     if (! $catid or ! isset($listcats[$catid])) {
@@ -43,12 +39,7 @@ if ($nv_Request->isset_request("catid", "get")) {
     }
 
     $caption = sprintf($lang_module['faq_list_by_cat'], $listcats[$catid]['title']);
-	if($where!='') {
-		$sql .= " WHERE ".$where." AND catid=" . $catid . " ORDER BY weight ASC";
-	}
-	else {
-		$sql .= " WHERE catid=" . $catid . " ORDER BY weight ASC";
-	}
+	$sql .= " WHERE catid=" . $catid . " ORDER BY weight ASC";
     $base_url .= "&amp;catid=" . $catid;
 
     define('NV_IS_CAT', true);
@@ -80,7 +71,7 @@ if ($all_page) {
 	        'id' => ( int )$row['id'], //
 	        'title' => $row['title'], //
 	        'cattitle' => $listcats[$row['catid']]['title'], //
-	        'link' => $link.'/#faq'.( int )$row['id'],
+	        'link' => $link.'#faq'.( int )$row['id'],
 	        );
 	}
 	$generate_page_accept = nv_generate_page($base_url, $all_page, $per_page, $page);
@@ -88,10 +79,6 @@ if ($all_page) {
 //list bài viết chưa được duyệt
 $sql = "SELECT SQL_CALC_FOUND_ROWS * FROM " . NV_PREFIXLANG . "_" . $module_data . "_tmp";
 $base_url = NV_BASE_SITEURL . "index.php?" . NV_LANG_VARIABLE . "=" . NV_LANG_DATA . "&" . NV_NAME_VARIABLE . "=" . $module_name."&". NV_OP_VARIABLE . "=list";
-if($user_info['userid']!=0) {
-	$where .="userid=".$user_info['userid'];
-}
-
 if ($nv_Request->isset_request("catid", "get")) {
     $catid = $nv_Request->get_int('catid', 'get', 0);
     if (! $catid or ! isset($listcats[$catid])) {
@@ -99,12 +86,7 @@ if ($nv_Request->isset_request("catid", "get")) {
         exit();
     }
     $caption = sprintf($lang_module['faq_list_by_cat'], $listcats[$catid]['title']);
-    if($where!='') {
-		$sql .= " WHERE ".$where." AND catid=" . $catid . " ORDER BY weight ASC";
-	}
-	else {
-		$sql .= " WHERE catid=" . $catid . " ORDER BY weight ASC";
-	}
+    $sql .= " WHERE catid=" . $catid . " ORDER BY weight ASC";
     $base_url .= "&amp;catid=" . $catid;
 
     define('NV_IS_CAT', true);
@@ -157,7 +139,7 @@ if ($nv_Request->isset_request('del', 'post')) {
         die('NO');
     }
 
-    $sql = "DELETE FROM " . NV_PREFIXLANG . "_" . $module_data . "_tmp WHERE id=" . $id;
+    $sql = "DELETE FROM " . NV_PREFIXLANG . "_" . $module_data . "_tmp WHERE id=" . $id ." AND userid=".$user_info['userid'];;
     $db->query($sql);
 
     nv_update_keywords($catid);
